backend/requirements.txt
Django==4.0.3
django-rest-framework==0.1.0
scikit-learn==1.0.2
pandas==1.4.1